+++ .storybook/vite.config.ts
@@ -1,9 +1,6 @@
-import path, { resolve } from "path";
-import { fileURLToPath } from "url";
+import { resolve } from "path";
 import react from "@vitejs/plugin-react";
 import { defineConfig } from "vite";
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const config = {
   assetsInclude: ["/sb-preview/runtime.js"],
@@ -14,7 +11,7 @@ const config = {
     alias: [
       {
         find: "@",
-        replacement: resolve(__dirname, "../src"),
+        replacement: resolve(__dirname, "src"),
       },
     ],
   },